refactor : 장소상세 지도부분 컴포넌트화(#197)
