+++ src/pages/place/[placeId]/index.tsx
@@ -6,18 +6,8 @@ import PlaceDetail from '@/components/place_details/PlaceDetailInfo';
 import { useRouter } from 'next/router';
 import Seo from '@/components/layout/Seo';
 import _ from 'lodash';
-import {
-  CustomOverlayRoadview,
-  Map,
-  MapMarker,
-  MapTypeControl,
-  Roadview,
-  RoadviewComponent,
-  RoadviewMarker,
-  ZoomControl,
-} from 'react-kakao-maps-sdk';
-import { useEffect, useRef, useState } from 'react';
-import { Button, Spinner } from '@nextui-org/react';
+import { useEffect, useState } from 'react';
+import { Spinner } from '@nextui-org/react';
 import CarouselThumb from '@/components/common/Carousel_Thumb';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/redux/config/configStore';
@@ -25,22 +15,21 @@ import { toastSuccess, toastWarn } from '@/libs/toastifyAlert';
 import PlaceDetailHeader from '@/components/place_details/PlaceDetailInfoHeader';
 import { getBookmark } from '@/apis/bookmarks';
 import { useBookmarks } from '@/hooks/useBookmarks';
-import { useCurrentTheme } from '@/hooks/useCurrentTheme';
 import PlaceDetailReview from '@/components/place_details/placeDetailReview';
+import PlaceDetailMap from '@/components/place_details/placeDetailMap';
 
 export type ToggleBookmarkType = () => void;
 export type ShowAlertType = () => void;
 
 const PlacePage = () => {
   const router = useRouter();
-  const roadviewRef = useRef<kakao.maps.Roadview | null>(null);
+
   const placeId: string = router.query.placeId as string;
-  const [toggle, setToggle] = useState('map');
+
   const [isBookmarked, setIsBookmarked] = useState<boolean>(false);
   const [recentOrder, setRecentOrder] = useState(true);
   const { userId, isLoggedIn } = useSelector((state: RootState) => state.auth);
   const { insertBookmark, deleteBookmark } = useBookmarks(userId, placeId);
-  const { baple } = useCurrentTheme();
 
   const { data: placeInfo, isLoading: placeInfoLoading } = useQuery({
     queryKey: ['placeInfo', placeId],
@@ -73,11 +62,6 @@ const PlacePage = () => {
     .flat()
     .filter((url) => url !== null) as string[];
 
-  let placePosition = {
-    lat: placeInfo?.lat,
-    lng: placeInfo?.lng,
-  };
-
   // 버튼 토글
   const toggleBookmark = () => {
     if (isBookmarked) {
@@ -108,15 +92,6 @@ const PlacePage = () => {
       </div>
     );
   }
-
-  const Content = () => {
-    return (
-      <div className='bg-white  text-black w-[250px] h-[100px] rounded-[20px] flex items-center justify-center text-[20px] font-bold text-wrap p-4'>
-        <p>{placeInfo.place_name}</p>
-        <div className='w-5 h-5 bg-white absolute bottom-[-10px] rotate-45'></div>
-      </div>
-    );
-  };
 
   return (
     <>
@@ -158,93 +133,7 @@ const PlacePage = () => {
         </div>
 
         {/* 지도 */}
-        <section className='mb-[30px] relative'>
-          {toggle === 'map' && (
-            <Map
-              center={placePosition}
-              draggable={false}
-              // zoomable={true}
-              scrollwheel={false}
-              keyboardShortcuts={true}
-              style={{
-                // 지도의 크기
-                width: '100%',
-                height: '300px',
-              }}
-              level={4}
-              minLevel={8}
-            >
-              <MapMarker
-                position={placePosition}
-                image={{
-                  src: `/images/icons/${
-                    baple ? 'marker.svg' : 'CBicons/CBmarker.svg'
-                  }`, // 마커이미지의 주소입니다
-                  size: {
-                    width: 44,
-                    height: 40,
-                  },
-                }}
-              />
-              <Button
-                className='absolute flex z-10 top-[3px] left-[130px] w-[90px] h-[32px] justify-center'
-                variant='solid'
-                color='primary'
-                onClick={() => setToggle('roadview')}
-                title='지도 보기'
-              >
-                로드뷰 보기
-              </Button>
-              <MapTypeControl position={'TOPLEFT'} />
-              <ZoomControl position={'LEFT'} />
-            </Map>
-          )}
-
-          {toggle === 'roadview' && (
-            <Roadview
-              position={{ ...placePosition, radius: 200 }}
-              style={{
-                width: '100%',
-                height: '300px',
-              }}
-              ref={roadviewRef}
-            >
-              {/* <RoadviewMarker position={placePosition} /> */}
-              <Button
-                className='absolute top-[5px] left-[5px] z-10 flex w-[90px] h-[32px] justify-center'
-                variant='solid'
-                color='primary'
-                onClick={() => setToggle('map')}
-                title='지도 보기'
-              >
-                지도
-              </Button>
-              <CustomOverlayRoadview
-                position={placePosition}
-                xAnchor={0.5}
-                yAnchor={0.5}
-                onCreate={(overlay) => {
-                  const roadview = roadviewRef.current;
-
-                  if (!roadview) {
-                    return;
-                  }
-
-                  const projection = roadview.getProjection(); // viewpoint(화면좌표)값을 추출할 수 있는 projection 객체를 가져옵니다.
-                  // 커스텀오버레이의 position과 altitude값을 통해 viewpoint값(화면좌표)를 추출합니다.
-                  console.log(overlay.getPosition());
-                  const viewpoint = projection.viewpointFromCoords(
-                    overlay.getPosition(),
-                    overlay.getAltitude(),
-                  );
-                  roadview.setViewpoint(viewpoint); //커스텀 오버레이를 로드뷰의 가운데에 오도록 로드뷰의 시점을 변화 시킵니다.
-                }}
-              >
-                <Content />
-              </CustomOverlayRoadview>
-            </Roadview>
-          )}
-        </section>
+        <PlaceDetailMap placeInfo={placeInfo} />
 
         {/* 리뷰 */}
         <PlaceDetailReview
